LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-range-animation-expected.txt


Harness Error (TIMEOUT), message = null

FAIL Animation with ranges [initial, initial] assert_equals: expected "0" but got "100"
TIMEOUT Animation with ranges [cover 0%, cover 100%] Test timed out
NOTRUN Animation with ranges [contain 0%, contain 100%]
NOTRUN Animation with ranges [entry 0%, entry 100%]
NOTRUN Animation with ranges [exit 0%, exit 100%]
NOTRUN Animation with ranges [contain -50%, entry 200%]
NOTRUN Animation with ranges [entry 0%, exit 100%]
NOTRUN Animation with ranges [cover 20px, cover 100px]
NOTRUN Animation with ranges [contain 20px, contain 100px]
NOTRUN Animation with ranges [entry 20px, entry 100px]
NOTRUN Animation with ranges [entry-crossing 20px, entry-crossing 100px]
NOTRUN Animation with ranges [exit 20px, exit 80px]
NOTRUN Animation with ranges [exit-crossing 20px, exit-crossing 80px]
NOTRUN Animation with ranges [contain 20px, contain calc(100px - 10%)]
NOTRUN Animation with ranges [exit 2em, exit 8em]
NOTRUN Animation with ranges [exit 2em, exit 8em] (scoped)
